requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
python-telegram-bot==21.1
sniffio==1.3.1
watchdog==4.0.0
yarl==1.9.4
